Added check for sharecode on_message events to prevent exceptions when the bot doesn't have the permission

+++ sharecode.py
@@ -85,10 +85,11 @@ class sharecode(commands.Cog):
 
     @commands.Cog.listener("on_message")
     async def on_message(self, message):
-        if isinstance(message.channel, discord.TextChannel):
+        if isinstance(message.channel, discord.TextChannel) and ctx.channel.permissions_for(ctx.guild.me).manage_webhooks:
             webhooks = await message.channel.webhooks()
             webhook = discord.utils.get(webhooks, name = self.bot.user.name)
             print(message)
+            print(webhook)
             if message.author.id != webhook.id:
                 guildDB = db.guilds.get(db.guilds.GuildID == message.guild.id)
                 channelDB = db.channels.get_or_create(ChannelID = message.channel.id)[0]
@@ -103,13 +104,15 @@ class sharecode(commands.Cog):
 
     @commands.Cog.listener("on_message_delete")
     async def on_message_delete(self, message):
-        tm = await transmitted().fetch(self.bot, message)
-        await tm.delete()
+        if ctx.channel.permissions_for(ctx.guild.me).manage_webhooks: 
+            tm = await transmitted().fetch(self.bot, message)
+            await tm.delete()
 
     @commands.Cog.listener("on_message_edit")
     async def on_message_edit(self, before, after):
-        tm = await transmitted().fetch(self.bot, before)
-        await tm.edit(content = after.content, embeds = after.embeds)
+        if not ctx.channel.permissions_for(ctx.guild.me).manage_webhooks: 
+            tm = await transmitted().fetch(self.bot, before)
+            await tm.edit(content = after.content, embeds = after.embeds)
             
 
     channel = discord.SlashCommandGroup("sharecode", locale_class.get('desc'))
